requirements.txt
azure-functions
psycopg2-binary
python-dotenv
facebook-business
requests